refactor: extract session queries into focused modules
Split lib/db/sessions.ts (1339 lines) into focused modules:
- lib/db/session-sets.ts: set CRUD operations (~300 lines)
- lib/db/session-stats.ts: analytics, PRs, volume queries (~450 lines)
- lib/db/exercise-history.ts: exercise history and records (~200 lines)
- lib/db/sessions.ts: reduced to session CRUD + re-exports (~250 lines)

All exports re-exported from sessions.ts for backward compatibility.
